@@ -28,14 +28,16 @@
     "test:appfrontend:headless": "cypress run -s 'e2e/integration/app-frontend/*.js'"
   },
   "devDependencies": {
-    "cypress": "^7.6.0",
+    "axe-core": "^4.3.2",
+    "cypress": "^8.3.0",
+    "cypress-axe": "^0.13.0",
     "cypress-file-upload": "^5.0.8",
     "cypress-plugin-tab": "^1.0.5",
-    "eslint": "^7.29.0",
+    "eslint": "^7.32.0",
     "eslint-plugin-cypress": "^2.11.3",
     "faker": "^5.5.3",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.3.2",
-    "start-server-and-test": "^1.12.5"
+    "start-server-and-test": "^1.13.1"
   }
 }
