@@ -167,7 +167,7 @@ const PartySelectionWithRouter = withRouter((props: IPartySelectionProps) => {
           // Keeping the switch statement because we might extends the enums to handle more errors
           case HttpStatusCodes.Forbidden: {
             return (
-              <Typography className={classes.partySelectionError}>
+              <Typography className={classes.partySelectionError} id='party-selection-error'>
                 {`
                   ${language.party_selection.invalid_selection_first_part} ${getRepresentedPartyName()}.
                   ${language.party_selection.invalid_selection_second_part} ${templatePartyTypesString()}.
