@@ -3,8 +3,8 @@ import { AppBar, createStyles, Grid, Typography, withStyles, WithStyles } from '
 import * as React from 'react';
 import { IParty } from '../../types';
 import { renderPartyName } from '../../utils/party';
-import AltinnIcon from '../AltinnIcon';
 import AltinnLogo from '../AltinnLogo';
+import AltinnAppHeaderMenu from './AltinnAppHeaderMenu';
 
 export interface IAltinnAppHeaderProps extends WithStyles<typeof styles> {
   /** The party of the instance owner */
@@ -15,6 +15,10 @@ export interface IAltinnAppHeaderProps extends WithStyles<typeof styles> {
   logoColor: string;
   /** The header background color */
   headerBackgroundColor: string;
+  /** The logout text */
+  logoutText: string;
+  /** The aria label text for profile menu */
+  ariaLabelIcon: string;
 }
 
 const styles = createStyles({
@@ -49,17 +53,19 @@ const styles = createStyles({
 });
 
 export function AltinnAppHeader(props: IAltinnAppHeaderProps) {
-  const {classes, logoColor, headerBackgroundColor: headerColor, party, userParty} = props;
+  const {
+    classes, logoColor, headerBackgroundColor: headerColor, party, userParty, logoutText, ariaLabelIcon,
+  } = props;
   return (
     <AppBar
-      position={'relative'}
-      classes={{root: classes.altinnAppHeader}}
-      style={{backgroundColor: headerColor, color: logoColor}}
+      position='relative'
+      classes={{ root: classes.altinnAppHeader }}
+      style={{ backgroundColor: headerColor, color: logoColor }}
     >
       <Grid
         container={true}
         className={classes.mainContent}
-        alignItems={'center'}
+        alignItems='center'
       >
         <Grid
           container={true}
@@ -67,13 +73,13 @@ export function AltinnAppHeader(props: IAltinnAppHeaderProps) {
           xs={6}
         >
           <Grid item={true}>
-            <AltinnLogo color={logoColor}/>
+            <AltinnLogo color={logoColor} />
           </Grid>
         </Grid>
         <Grid
           container={true}
-          justifyContent={'flex-end'}
-          alignItems={'center'}
+          justifyContent='flex-end'
+          alignItems='center'
           item={true}
           xs={6}
         >
@@ -84,7 +90,10 @@ export function AltinnAppHeader(props: IAltinnAppHeaderProps) {
               </Typography>
             }
             {(party && userParty && party.partyId !== userParty.partyId) &&
-              <Grid container={true} direction={'column'} alignItems={'flex-end'}>
+              <Grid
+                container={true} direction='column'
+                alignItems='flex-end'
+              >
                 <Grid item={true}>
                   <Typography className={classes.appHeaderText}>
                     {renderPartyName(userParty)}
@@ -99,22 +108,12 @@ export function AltinnAppHeader(props: IAltinnAppHeaderProps) {
             }
           </Grid>
           <Grid item={true}>
-            {party && party.ssn &&
-              <AltinnIcon
-                iconClass={'fa fa-private-circle-big'}
-                iconColor={logoColor}
-                iconSize={31}
-                margin={'0px 0px 0px 5px'}
-              />
-            }
-            {party && party.orgNumber &&
-              <AltinnIcon
-                iconClass={'fa fa-corp-circle-big'}
-                iconColor={logoColor}
-                iconSize={31}
-                margin={'0px 0px 0px 5px'}
-              />
-            }
+            <AltinnAppHeaderMenu
+              party={party}
+              logoColor={logoColor}
+              logoutText={logoutText}
+              ariaLabel={ariaLabelIcon}
+            />
           </Grid>
         </Grid>
       </Grid>
