@@ -75,31 +75,33 @@ export function en() {
     },
     general: {
       action: 'Action',
-      add: 'Add',
       add_connection: 'Add connection',
       add_new: 'Add new',
+      add: 'Add',
       back: 'Back',
       cancel: 'Cancel',
-      choose: 'Choose',
       choose_label: 'Choose label',
       choose_method: 'Choose method',
-      close: 'Close',
+      choose: 'Choose',
       close_schema: 'Close schema',
+      close: 'Close',
       contains: 'Contains{0}',
       control_submit: 'Control and submit',
-      create: 'Create',
       create_new: 'Create new',
+      create: 'Create',
       customer_service_phone_number: '+47 75 00 60 00',
       delete: 'Delete',
       disabled: 'Disabled',
-      edit: 'Edit',
-      edit_alt: 'Edit',
       edit_alt_error: 'Fix error here',
+      edit_alt: 'Edit',
+      edit: 'Edit',
       enabled: 'Enabled',
       error_message_with_colon: 'Error message:',
       for: 'for',
+      header_profile_icon_label: 'Profile icon button',
       label: 'Label',
       loading: 'Loading content',
+      log_out: 'Log out',
       next: 'Next',
       no_options: 'No options available',
       optional: 'Optional',
