@@ -23,17 +23,6 @@ namespace Altinn.Platform.Authentication.Configuration
         public string AltinnPartyCookieName { get; set; }
 
         /// <summary>
-        /// Gets the altinnParty cookie from kubernetes environment variables and appsettings if environment variable is not set
-        /// </summary>
-        public string GetAltinnPartyCookieName
-        {
-            get
-            {
-                return Environment.GetEnvironmentVariable("GeneralSettings__AltinnPartyCookieName") ?? AltinnPartyCookieName;
-            }
-        }
-
-        /// <summary>
         /// Gets or sets the bridge authentication api endpoint
         /// </summary>
         public string BridgeAuthnApiEndpoint { get; set; }
@@ -49,17 +38,6 @@ namespace Altinn.Platform.Authentication.Configuration
         public string SBLRedirectEndpoint { get; set; }
 
         /// <summary>
-        /// Gets the sbl redirect endpoint from kubernetes environment variables and appsettings if environment variable is not set
-        /// </summary>
-        public string GetSBLRedirectEndpoint
-        {
-            get
-            {
-                return Environment.GetEnvironmentVariable("GeneralSettings__SBLRedirectEndpoint") ?? SBLRedirectEndpoint;
-            }
-        }
-
-        /// <summary>
         /// Gets or sets the platform endpoint
         /// </summary>
         public string PlatformEndpoint { get; set; }
@@ -68,17 +46,6 @@ namespace Altinn.Platform.Authentication.Configuration
         /// Gets or sets the claims identity
         /// </summary>
         public string ClaimsIdentity { get; set; }
-
-        /// <summary>
-        /// Gets the claims identity from kubernetes environment variables and appsettings if environment variable is not set
-        /// </summary>
-        public string GetClaimsIdentity
-        {
-            get
-            {
-                return Environment.GetEnvironmentVariable("GeneralSettings__ClaimsIdentity") ?? ClaimsIdentity;
-            }
-        }
 
         /// <summary>
         /// Gets or sets the number of minutes the JSON Web Token and the cookie is valid.
@@ -91,48 +58,14 @@ namespace Altinn.Platform.Authentication.Configuration
         public string HostName { get; set; }
 
         /// <summary>
-        /// Gets the jwt cookie validity time from kubernetes environment variables and appsettings if environment variable is not set
-        /// </summary>
-        public string GetHostName
-        {
-            get
-            {
-                return Environment.GetEnvironmentVariable("GeneralSettings__HostName") ?? HostName;
-            }
-        }
-
-        /// <summary>
         /// Gets or sets the BaseUrl
         /// </summary>
         public string BaseUrl { get; set; }
 
         /// <summary>
-        /// Gets the jwt cookie validity time from kubernetes environment variables and appsettings if environment variable is not set
-        /// </summary>
-        public string GetBaseUrl
-        {
-            get
-            {
-                return Environment.GetEnvironmentVariable("GeneralSettings__BaseUrl") ?? BaseUrl;
-            }
-        }
-
-        /// <summary>
         /// Gets or sets URL of the well known configuration endpoint for Maskinporten.
         /// </summary>
         public string MaskinportenWellKnownConfigEndpoint { get; set; }
-
-        /// <summary>
-        /// Gets or sets URL of the well known configuration endpoint for Maskinporten from kubernetes or appsettings if no environment variable is set.
-        /// </summary>
-        public string GetMaskinportenWellKnownConfigEndpoint
-        {
-            get
-            {
-                return Environment.GetEnvironmentVariable("GeneralSettings__" + nameof(MaskinportenWellKnownConfigEndpoint)) ??
-                       MaskinportenWellKnownConfigEndpoint;
-            }
-        }
 
         /// <summary>
         /// Gets url of the well known configuration endpoint for ID-porten from environment variable.
@@ -143,18 +76,6 @@ namespace Altinn.Platform.Authentication.Configuration
         /// Gets or sets the url to the json file which holds the valid organisation entries (which inclides name, organisation number and org identifier)
         /// </summary>
         public string OrganisationRepositoryLocation { get; set; }
-
-        /// <summary>
-        /// Gets the url of the list of valid organisation entries (json)
-        /// </summary>
-        public string GetOrganisationRepositoryLocation
-        {
-            get
-            {
-                return Environment.GetEnvironmentVariable("GeneralSettings__" + nameof(OrganisationRepositoryLocation)) ??
-                    OrganisationRepositoryLocation;
-            }
-        }
 
         /// <summary>
         /// Gets or sets the URL of the Altinn Open ID Connect well-known configuration endpoint.
@@ -171,6 +92,17 @@ namespace Altinn.Platform.Authentication.Configuration
         /// Take care not to upload "old" certificates.
         /// </remarks>
         public int JwtSigningCertificateRolloverDelayHours { get; set; }
+
+        /// <summary>
+        /// Gets the Altinn Open ID Connect (OIDC) Issuer URL.
+        /// </summary>
+        public string AltinnOidcIssuerUrl
+        {
+            get
+            {
+                return PlatformEndpoint + "authentication/api/v1/openid/";
+            }
+        }
 
         /// <summary>
         /// Get or sets the value indicating if OIDC authentication is enabled
