@@ -46,7 +46,7 @@ namespace Altinn.Platform.Authentication.Controllers
         [HttpGet("openid-configuration")]
         public async Task<IActionResult> GetOpenIdConfigurationAsync()
         {
-            string baseUrl = _generalSettings.PlatformEndpoint;
+            string baseUrl = _generalSettings.AltinnOidcIssuerUrl;
 
             DiscoveryDocument discoveryDocument = new DiscoveryDocument
             {
@@ -60,7 +60,7 @@ namespace Altinn.Platform.Authentication.Controllers
                 TokenEndpoint = new Uri(baseUrl).ToString(),
 
                 // REQUIRED
-                JwksUri = new Uri(baseUrl + "authentication/api/v1/OpenId/.well-known/openid-configuration/jwks").ToString(),
+                JwksUri = new Uri(baseUrl + ".well-known/openid-configuration/jwks").ToString(),
 
                 // REQUIRED
                 ResponseTypesSupported = new[] { "token" }, // "code", "id_token", "id_token token", 
