@@ -1,9 +1,12 @@
 using System.Linq;
 using System.Threading.Tasks;
+
 using Altinn.Platform.Profile.Models;
 using Altinn.Platform.Profile.Services.Interfaces;
 using AltinnCore.Authentication.Constants;
+
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Altinn.Platform.Profile.Controllers
@@ -12,7 +15,9 @@ namespace Altinn.Platform.Profile.Controllers
     /// The users controller
     /// </summary>
     [Authorize]
-    [Route("profile/api/v1/[controller]")]
+    [Route("profile/api/v1/users")]
+    [Consumes("application/json")]
+    [Produces("application/json")]
     public class UsersController : Controller
     {
         private readonly IUserProfiles _userProfilesWrapper;
@@ -33,7 +38,9 @@ namespace Altinn.Platform.Profile.Controllers
         /// <returns>The information about a given user</returns>
         [HttpGet("{userID}")]
         [Authorize(Policy = "PlatformAccess")]
-        public async Task<ActionResult> Get(int userID)
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        public async Task<ActionResult<UserProfile>> Get(int userID)
         {
             UserProfile result = await _userProfilesWrapper.GetUser(userID);
             if (result == null)
@@ -49,10 +56,13 @@ namespace Altinn.Platform.Profile.Controllers
         /// </summary>
         /// <returns>User profile of current user</returns>
         [HttpGet("current")]
-        public async Task<ActionResult> Get()
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        public async Task<ActionResult<UserProfile>> Get()
         {
-            string userIdString = Request.HttpContext.User.Claims.Where(c => c.Type == AltinnCoreClaimTypes.UserId)
-           .Select(c => c.Value).SingleOrDefault();
+            string userIdString = Request.HttpContext.User.Claims
+                .Where(c => c.Type == AltinnCoreClaimTypes.UserId)
+                .Select(c => c.Value).SingleOrDefault();
 
             if (string.IsNullOrEmpty(userIdString))
             {
@@ -71,7 +81,9 @@ namespace Altinn.Platform.Profile.Controllers
         /// <returns>User profile connected to given SSN </returns>
         [HttpPost]
         [Authorize(Policy = "PlatformAccess")]
-        public async Task<ActionResult> GetUserFromSSN([FromBody]string ssn)
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        public async Task<ActionResult<UserProfile>> GetUserFromSSN([FromBody]string ssn)
         {
             UserProfile result = await _userProfilesWrapper.GetUser(ssn);
             if (result == null)
