@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Net;
 using System.Net.Http;
+using System.Net.Http.Headers;
 using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
@@ -356,7 +357,7 @@ namespace Designer.Tests.Controllers
             var developer = "testUser";
             var targetRepository = Guid.NewGuid().ToString();
 
-            await TestDataHelper.CopyAppRepositoryForTest(org, sourceRepository, developer, targetRepository);
+            await TestDataHelper.CopyRepositoryForTest(org, sourceRepository, developer, targetRepository);
 
             var client = GetTestClient();
             var url = $"{_versionPrefix}/{org}/{targetRepository}/Datamodels/?modelPath={modelPath}";
@@ -373,6 +374,43 @@ namespace Designer.Tests.Controllers
                 var response = await client.SendAsync(httpRequestMessage);
 
                 Assert.Equal(HttpStatusCode.NoContent, response.StatusCode);
+            }
+            finally
+            {
+                TestDataHelper.DeleteAppRepository(org, targetRepository, developer);
+            }
+        }
+
+        [Fact]
+        public async Task PostDatamodel_FromXsd_ShouldReturnCreated()
+        {
+            // Arrange
+            var org = "ttd";
+            var sourceRepository = "empty-datamodels";
+            var developer = "testUser";
+            var targetRepository = Guid.NewGuid().ToString();
+
+            await TestDataHelper.CopyRepositoryForTest(org, sourceRepository, developer, targetRepository);
+            var client = GetTestClient();
+            var url = $"{_versionPrefix}/{org}/{targetRepository}/Datamodels";
+
+            var fileStream = TestDataHelper.LoadDataFromEmbeddedResource("Designer.Tests._TestData.Model.Xsd.Kursdomene_HvemErHvem_M_2021-04-08_5742_34627_SERES.xsd");
+            var formData = new MultipartFormDataContent();
+            var streamContent = new StreamContent(fileStream);
+            streamContent.Headers.ContentType = MediaTypeHeaderValue.Parse("multipart/form-data");
+            formData.Add(streamContent, "file", "Kursdomene_HvemErHvem_M_2021-04-08_5742_34627_SERES.xsd");
+
+            var httpRequestMessage = new HttpRequestMessage(HttpMethod.Post, url)
+            {
+                Content = formData
+            };
+
+            await AuthenticationUtil.AddAuthenticateAndAuthAndXsrFCookieToRequest(client, httpRequestMessage);
+
+            try
+            {
+                var response = await client.SendAsync(httpRequestMessage);
+                Assert.Equal(HttpStatusCode.Created, response.StatusCode);
             }
             finally
             {
