@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
@@ -10,15 +9,10 @@ using System.Xml.Schema;
 using System.Xml.Serialization;
 using Altinn.Studio.Designer.Factories.ModelFactory;
 using Altinn.Studio.Designer.ModelMetadatalModels;
-using Basic.Reference.Assemblies;
 using Designer.Tests.Utils;
 using FluentAssertions;
 using Manatee.Json;
 using Manatee.Json.Schema;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.Emit;
-using Microsoft.CodeAnalysis.Text;
 using Xunit;
 using Xunit.Abstractions;
 using XmlSchemaValidator = Designer.Tests.Utils.XmlSchemaValidator;
@@ -45,7 +39,7 @@ namespace Designer.Tests.Factories.ModelFactory
             ModelMetadata modelMetadata = GenerateModelMetadata(org, app, jsonSchema);
             string classes = GenerateCSharpClasses(modelMetadata);
                         
-            Assembly assembly = CompileToAssembly(classes);            
+            Assembly assembly = Compiler.CompileToAssembly(classes);            
             
             Type type = assembly.GetType("Altinn.App.Models.melding");
 
@@ -95,7 +89,7 @@ namespace Designer.Tests.Factories.ModelFactory
             ModelMetadata modelMetadata = GenerateModelMetadata(org, app, jsonSchema);
             string classes = GenerateCSharpClasses(modelMetadata);
 
-            Assembly assembly = CompileToAssembly(classes);
+            Assembly assembly = Compiler.CompileToAssembly(classes);
 
             Type type = assembly.GetType(modelName);
 
@@ -153,12 +147,12 @@ namespace Designer.Tests.Factories.ModelFactory
             ModelMetadata modelMetadata = GenerateModelMetadata(org, app, jsonSchema);
 
             string classes = GenerateCSharpClasses(modelMetadata);
-            Assembly assembly = CompileToAssembly(classes);
+            Assembly assembly = Compiler.CompileToAssembly(classes);
             Type type = assembly.GetType(modelName);
             var modelInstance = assembly.CreateInstance(type.FullName);
 
             string expectedClasses = TestDataHelper.LoadDataFromEmbeddedResourceAsString(expectedCSharpResource);
-            Assembly expectedAssembly = CompileToAssembly(expectedClasses);
+            Assembly expectedAssembly = Compiler.CompileToAssembly(expectedClasses);
             Type expectedType = expectedAssembly.GetType(modelName);
             var expectedModelInstance = expectedAssembly.CreateInstance(expectedType.FullName);
             expectedType.HasSameMetadataDefinitionAs(type);
@@ -191,45 +185,6 @@ namespace Designer.Tests.Factories.ModelFactory
             JsonMetadataParser modelGenerator = new JsonMetadataParser();
             string classes = modelGenerator.CreateModelFromMetadata(modelMetadata);
             return classes;
-        }
-
-        private Assembly CompileToAssembly(string classes)
-        {
-            var syntaxTree = SyntaxFactory.ParseSyntaxTree(SourceText.From(classes));
-
-            var compilation = CSharpCompilation.Create(Guid.NewGuid().ToString())
-                .WithOptions(new CSharpCompilationOptions(OutputKind.DynamicallyLinkedLibrary))
-                .WithReferenceAssemblies(ReferenceAssemblyKind.Net50)
-                .AddReferences(MetadataReference.CreateFromFile(typeof(Microsoft.AspNetCore.Mvc.ModelBinding.BindNeverAttribute).GetTypeInfo().Assembly.Location))
-                .AddReferences(MetadataReference.CreateFromFile(typeof(Newtonsoft.Json.JsonPropertyAttribute).GetTypeInfo().Assembly.Location))
-                .AddSyntaxTrees(syntaxTree);
-
-            Assembly assembly = null;
-            using (var ms = new MemoryStream())
-            {
-                EmitResult result = compilation.Emit(ms);
-
-                Assert.True(result.Success);
-
-                if (!result.Success)
-                {
-                    IEnumerable<Diagnostic> failures = result.Diagnostics.Where(diagnostic =>
-                        diagnostic.IsWarningAsError ||
-                        diagnostic.Severity == DiagnosticSeverity.Error);
-
-                    foreach (Diagnostic diagnostic in failures)
-                    {
-                        _outputHelper.WriteLine("{0}: {1}", diagnostic.Id, diagnostic.GetMessage());                        
-                    }
-                }
-                else
-                {
-                    ms.Seek(0, SeekOrigin.Begin);
-                    assembly = Assembly.Load(ms.ToArray());
-                }
-            }
-
-            return assembly;
         }
     }
 }
