@@ -1,7 +1,9 @@
 using System;
 using System.IO;
 using System.Threading.Tasks;
+
 using Designer.Tests.Utils;
+
 using Xunit;
 
 namespace Designer.Tests.Infrastructure.GitRepository
@@ -13,7 +15,7 @@ namespace Designer.Tests.Infrastructure.GitRepository
         [InlineData(@"This should be read back with the same formatting!\n")]
         [InlineData(@"{""some"":""random"", ""json"":""value""}")]
         public async Task WriteTextByRelativePathAsync_ValidText_ShouldReadBackEqual(string expectedContent)
-        {   
+        {
             string repositoriesRootDirectory = TestDataHelper.GetTestDataRepositoriesRootDirectory();
             string repositoryDirectory = TestDataHelper.GetTestDataRepositoryDirectory("ttd", "hvem-er-hvem", "testUser");
             var gitRepository = new Altinn.Studio.Designer.Infrastructure.GitRepository.GitRepository(repositoriesRootDirectory, repositoryDirectory);
@@ -39,7 +41,7 @@ namespace Designer.Tests.Infrastructure.GitRepository
         [InlineData(@"app/models/HvemErHvem_SERES.schema.json")]
         [InlineData(@"/app/models/HvemErHvem_SERES.schema.json")]
         public async Task WriteTextByRelativePathAsync_ReadWriteRoundtrip_ShouldReadBackEqual(string expectedFilePath)
-        {            
+        {
             var gitRepository = GetTestRepository("ttd", "hvem-er-hvem", "testUser");
 
             var expectedContent = await gitRepository.ReadTextByRelativePathAsync(expectedFilePath);
@@ -48,7 +50,7 @@ namespace Designer.Tests.Infrastructure.GitRepository
             try
             {
                 await gitRepository.WriteTextByRelativePathAsync(filename, expectedContent);
-                var actualContent = await gitRepository.ReadTextByRelativePathAsync(filename); 
+                var actualContent = await gitRepository.ReadTextByRelativePathAsync(filename);
 
                 Assert.Equal(expectedContent, actualContent);
             }
@@ -89,15 +91,17 @@ namespace Designer.Tests.Infrastructure.GitRepository
 
             Assert.True(gitRepository.FileExistsByRelativePath(relativeFileUrl));
 
-            TestDataHelper.DeleteDirectory(Path.Join(gitRepository.RepositoryDirectory, "test_directory"));            
+            TestDataHelper.DeleteDirectory(Path.Join(gitRepository.RepositoryDirectory, "test_directory"));
         }
 
-        [Fact]
-        public void FileExistsByRelativePath_FileDontExits_ShouldReturnFalse()
+        [Theory]
+        [InlineData(@"this.dont.exists.schema.json")]
+        [InlineData(@"c:/this/should/not/exist/HvemErHvem.json")]
+        public void FileExistsByRelativePath_FileDontExits_ShouldReturnFalse(string relativePath)
         {
             var gitRepository = GetTestRepository("ttd", "hvem-er-hvem", "testUser");
 
-            Assert.False(gitRepository.FileExistsByRelativePath("this.dont.exists.schema.json"));
+            Assert.False(gitRepository.FileExistsByRelativePath(relativePath));
         }
 
         [Theory]
@@ -110,6 +114,63 @@ namespace Designer.Tests.Infrastructure.GitRepository
             var gitRepository = GetTestRepository("ttd", "hvem-er-hvem", "testUser");
 
             Assert.True(gitRepository.FileExistsByRelativePath(relativePath));
+        }
+
+        [Fact]
+        public void DirectoryExistsByRelativePath_Directory_ShouldReturnFalse()
+        {
+            var gitRepository = GetTestRepository("ttd", "hvem-er-hvem", "testUser");
+
+            Assert.False(gitRepository.DirectoryExitsByRelativePath("c:/this/does/not/exists"));
+        }
+
+        [Fact]
+        public void CopyRepository_DirectoryAlreadyExists_AllFilesSuccessfullyCopied()
+        {
+            // Arrange
+            string targetPath = TestDataHelper.CreateEmptyDirectory("cloneDirectory");
+            var gitRepository = GetTestRepository("ttd", "hvem-er-hvem", "testUser");
+
+            try
+            {
+                // Act
+                gitRepository.CopyRepository(targetPath);
+                int actualFileCount = Directory.GetFiles(targetPath, "*", SearchOption.AllDirectories).Length;
+
+                // Assert
+                int expectedFileCount = Directory.GetFiles(gitRepository.RepositoryDirectory, "*", SearchOption.AllDirectories).Length;
+
+                Assert.Equal(expectedFileCount, actualFileCount);
+            }
+            finally
+            {
+                Directory.Delete(targetPath, true);
+            }
+        }
+
+        [Fact]
+        public void CopyRepository_DirDoestNotExists_DirCreatedAndFilesCopied()
+        {
+            // Arrange
+            string targetPath = $"{TestDataHelper.GetTestDataRepositoriesRootDirectory()}/newClonedApp";
+            var gitRepository = GetTestRepository("ttd", "hvem-er-hvem", "testUser");
+
+            try
+            {
+                // Act
+                gitRepository.CopyRepository(targetPath);
+                int actualFileCount = Directory.GetFiles(targetPath, "*", SearchOption.AllDirectories).Length;
+
+                // Assert
+                int expectedFileCount = Directory.GetFiles(gitRepository.RepositoryDirectory, "*", SearchOption.AllDirectories).Length;
+
+                Assert.True(Directory.Exists(targetPath));
+                Assert.Equal(expectedFileCount, actualFileCount);
+            }
+            finally
+            {
+                Directory.Delete(targetPath, true);
+            }
         }
 
         private static Altinn.Studio.Designer.Infrastructure.GitRepository.GitRepository GetTestRepository(string org, string repository, string developer)
