@@ -1,6 +1,6 @@
 using System;
-using System.Collections.Generic;
 using System.IO;
+using System.Text.RegularExpressions;
 
 namespace Altinn.Studio.Designer.Helpers
 {
@@ -85,6 +85,32 @@ namespace Altinn.Studio.Designer.Helpers
             if (!Directory.Exists(directoryPath))
             {
                 throw new DirectoryNotFoundException($"Could not find the specified directory at '{directoryPath}'");
+            }
+        }
+
+        /// <summary>
+        /// Assert that a file has a specific extension.
+        /// </summary>
+        /// <param name="fileName">The fileName, full path, relative path or just the name, to verify.</param>
+        /// <param name="fileExtension">The extension the filename should have including the leading . e.g. '.xsd'.</param>
+        public static void AssertFileExtensionIsOfType(string fileName, string fileExtension)
+        {
+            var fileInfo = new FileInfo(fileName);
+            if (fileInfo.Extension.ToLower() != fileExtension.ToLower())
+            {
+                throw new ArgumentException($"The file {fileName} must be of type {fileExtension}.");
+            }
+        }
+
+        /// <summary>
+        /// Assert that a repository name is valid.
+        /// </summary>
+        /// <param name="repoName">The repository name.</param>
+        public static void AssertValidAppRepoName(string repoName)
+        {
+            if (string.IsNullOrEmpty(repoName) || !Regex.IsMatch(repoName, @"^(?!datamodels$)[a-z]+[a-z0-9-]+[a-z0-9]$"))
+            {
+                throw new ArgumentException($"The repository name {repoName} is invalid.");
             }
         }
     }
