@@ -34,25 +34,8 @@
 
   @RenderSection("styles", required: false)
   <style>
-    @@font-face {
-      font-family: "AltinnRoboto";
-      font-weight: bold;
-      src: url("/designer/font/roboto/roboto-bold-altinn.woff") format('woff');
-    }
-
-    @@font-face {
-      font-family: "AltinnRoboto";
-      font-weight: 100;
-      src: url("/designer/font/roboto/roboto-light-altinn.woff") format('woff');
-    }
-
-    @@font-face {
-      font-family: "AltinnRoboto";
-      src: url("/designer/font/roboto/roboto-regular-altinn.woff") format('woff');
-    }
-
     body {
-      font-family: AltinnRoboto, "San Fransisco", "Helvetica Neue", Helvetica, Arial, sans-serif;
+      font-family: Roboto, "San Fransisco", "Helvetica Neue", Helvetica, Arial, sans-serif;
     }
 
     h1, h2, h3 {
