@@ -1,15 +1,11 @@
 /* tslint:disable:jsx-wrap-multiline */
-// import { mount } from 'enzyme';
 import 'jest';
 import * as React from 'react';
 import { Provider } from 'react-redux';
 import * as renderer from 'react-test-renderer';
 import configureStore from 'redux-mock-store';
-import {
-  getListOfServicesExcludingCodelist,
-  ServicesOverview,
-  // ServicesOverviewComponent,
-} from '../../../features/serviceOverview/servicesOverview';
+import { IRepository } from '../../../../shared/types';
+import { getListOfServicesExcludingDatamodels, ServicesOverview } from '../../../features/serviceOverview/servicesOverview';
 
 describe('>>> features/serviceOverview', () => {
   let mockStore: any;
@@ -18,7 +14,9 @@ describe('>>> features/serviceOverview', () => {
     const createStore = configureStore();
     const initialState = {
       dashboard: {
-        user: 'Ulf Utvikler',
+        user: {
+          full_name: 'Ulf Utvikler',
+        },
         services: [
           {
             name: 'testService',
@@ -51,65 +49,47 @@ describe('>>> features/serviceOverview', () => {
     };
 
     mockStore = createStore(initialState);
-
   });
 
   it('>>> Capture snapshot of serviceOverview', () => {
     const rendered = renderer.create(
       <Provider store={mockStore}>
-         <ServicesOverview />
+        <ServicesOverview />
       </Provider>,
     );
     expect(rendered).toMatchSnapshot();
   });
 
-  // TODO: Removed failing test related to Material UI v4 upgrade. Should be rewritten in Jest or Test Café
-  // Test does not actually verify that the rendered result is displayed, only that the function is executed.
-  // it('+++ should run searchAndFilterServicesIntoCategoriesCategory on search', () => {
-  //   const mountedServiceOverview = mount(
-  //     <Provider store={mockStore}>
-  //        <ServicesOverview />
-  //     </Provider>,
-  //   );
-  //   const instance = mountedServiceOverview.find('ServicesOverviewComponent').instance() as ServicesOverviewComponent;
-  //   const spy = jest.spyOn(instance, 'searchAndFilterServicesIntoCategoriesCategory');
-
-  //   const searchField = mountedServiceOverview.find('input');
-  //   searchField.simulate('change', { target: {value: 'test'}});
-  //   expect(spy).toHaveBeenCalled();
-  // });
-
-  it('+++ if there are no services getListOfServicesExcludingCodelist should return null', () => {
-    const services = getListOfServicesExcludingCodelist(null);
-
-    expect(services).toEqual(null);
+  it('+++ if there are no services getListOfServicesExcludingDatamodels should return null', () => {
+    const services = getListOfServicesExcludingDatamodels(null);
+    expect(services).toEqual(undefined);
   });
 
-  it('+++ if there are services getListOfServicesExcludingCodelist should return services without codelists', () => {
-    const serviceList = [
+  it('+++ if there are services getListOfServicesExcludingDatamodels should return services without datamodels', () => {
+    const serviceList: IRepository[] = [
       {
         name: 'testService',
         owner: { full_name: 'Ulf Utvikler' },
         permissions: {
           push: true,
         },
-      },
+      } as unknown as IRepository,
       {
         name: 'NullSkatt',
         owner: { full_name: 'Ulf Utvikler' },
         permissions: {
           push: true,
         },
-      },
+      } as unknown as IRepository,
       {
-        name: 'codelists',
+        name: 'test-datamodels',
         owner: { full_name: 'Ulf Utvikler' },
         permissions: {
           push: true,
         },
-      },
+      } as unknown as IRepository,
     ];
-    const services = getListOfServicesExcludingCodelist(serviceList);
+    const services = getListOfServicesExcludingDatamodels(serviceList);
     const mockResult = [
       {
         name: 'testService',
