@@ -1,4 +1,4 @@
 import { configureStore } from '@reduxjs/toolkit';
 import schemaReducer from './features/editor/schemaEditorSlice';
 
-export const store = configureStore({reducer: schemaReducer});
+export const store = configureStore({ reducer: schemaReducer });
